+++ src/plus/integrations/providers/github/githubGitProvider.ts
@@ -49,7 +49,6 @@ import type { GitFile } from '../../../../git/models/file';
 import { GitFileChange, GitFileIndexStatus } from '../../../../git/models/file';
 import type {
 	GitGraph,
-	GitGraphHostingServiceType,
 	GitGraphRow,
 	GitGraphRowContexts,
 	GitGraphRowHead,
@@ -1424,7 +1423,7 @@ export class GitHubGitProvider implements GitProvider, Disposable {
 							avatarUrl: avatarUrl,
 							context: serializeWebviewItemContext<GraphItemRefContext>(context),
 							current: true,
-							hostingServiceType: remote.provider?.id as GitGraphHostingServiceType,
+							hostingServiceType: remote.provider?.gkProviderId,
 						},
 					];
 
@@ -1476,7 +1475,7 @@ export class GitHubGitProvider implements GitProvider, Disposable {
 							url: remote.url,
 							avatarUrl: avatarUrl,
 							context: serializeWebviewItemContext<GraphItemRefContext>(context),
-							hostingServiceType: remote.provider?.id as GitGraphHostingServiceType,
+							hostingServiceType: remote.provider?.gkProviderId,
 						});
 					}
 				}